Add files via upload
OBJECTIVE :

The project entitled "DRIVER DROWSINESS DETECTION" is a safety 
technology that helps prevent accidents caused by the driver getting drowsy. The 
number of road accidents due to Drowsy Driving is increasing at an alarming
rate worldwide.
• With the advancement in Computer Vision technologies, smart/intelligent
cameras are developed to identify drowsiness in drivers, thereby alerting drivers
which in turn reduces accidents when they are in fatigue.
• In this mini project, a new framework is proposed using deep learning to detect
driver drowsiness based on Eye state while driving the vehicle. To detect the face 
and extract the eye region from the face images, the Viola-Jones face detection
algorithm is used.
• Deep Convolution Neural Network is developed to extract features from
dynamically identified keyframes from camera sequences and used for the 
learning phase. A SoftMax layer in the CNN classifier is used to classify the
driver as sleep or non-sleep.
• This system alerts the driver with an alarm when the driver is in a sleepy mood.
The proposed work is evaluated on a collected dataset and shows better
accuracy

TOOL DESCRIPTION :
-Spyder
-Python

ALGORITHM :
-Convolutional Neural Network
-Viola Jones Algorithm